Rename for generalization and naming purposes
